Wrap MCP servers to lazy load items to not block the UI Mobile bug fixes
