+++ frontend/src/pages/Admin/Agents/Imported/SkillList/index.jsx
@@ -1,5 +1,4 @@
 import { CaretRight } from "@phosphor-icons/react";
-import { isMobile } from "react-device-detect";
 import { sentenceCase } from "text-case";
 
 export default function ImportedSkillList({
@@ -27,9 +26,7 @@ export default function ImportedSkillList({
 
   return (
     <div
-      className={`bg-theme-bg-secondary text-white rounded-xl ${
-        isMobile ? "w-full" : "min-w-[360px] w-fit"
-      }`}
+      className={`bg-theme-bg-secondary text-white rounded-xl w-full md:min-w-[360px]`}
     >
       {skills.map((config, index) => (
         <div
