+++ frontend/src/pages/Admin/Agents/MCPServers/index.jsx
@@ -1,11 +1,102 @@
-import React from "react";
+import { useState, useEffect } from "react";
 import { titleCase } from "text-case";
+import { Link } from "react-router-dom";
+import { BookOpenText, ArrowClockwise } from "@phosphor-icons/react";
+import MCPLogo from "@/media/agents/mcp-logo.svg";
+import MCPServers from "@/models/mcpServers";
 
-export default function MCPServersList({
+export function MCPServerHeader({ setMcpServers, children }) {
+  const [loadingMcpServers, setLoadingMcpServers] = useState(false);
+  useEffect(() => {
+    async function fetchMCPServers() {
+      setLoadingMcpServers(true);
+      const { servers = [] } = await MCPServers.listServers();
+      setMcpServers(servers);
+      setLoadingMcpServers(false);
+    }
+    fetchMCPServers();
+  }, []);
+
+  // Refresh the list of MCP servers
+  const refreshMCPServers = () => {
+    if (
+      window.confirm(
+        "Are you sure you want to refresh the list of MCP servers? This will restart all MCP servers and reload their tools."
+      )
+    ) {
+      setLoadingMcpServers(true);
+      MCPServers.forceReload()
+        .then(({ servers = [] }) => {
+          setSelectedMcpServer(null);
+          setMcpServers(servers);
+        })
+        .catch((err) => {
+          console.error(err);
+          showToast(`Failed to refresh MCP servers.`, "error", { clear: true });
+        })
+        .finally(() => {
+          setLoadingMcpServers(false);
+        });
+    }
+  };
+
+  return (
+    <>
+      <div className="text-theme-text-primary flex items-center justify-between gap-x-2 mt-4">
+        <div className="flex items-center gap-x-2">
+          <img src={MCPLogo} className="w-6 h-6 light:invert" />
+          <p className="text-lg font-medium">MCP Servers</p>
+        </div>
+        <div className="flex items-center gap-x-3">
+          <Link
+            to="#goes-to-docs"
+            target="_blank"
+            className="border-none text-theme-text-secondary hover:text-cta-button"
+          >
+            <BookOpenText size={16} />
+          </Link>
+          <button
+            type="button"
+            onClick={refreshMCPServers}
+            disabled={loadingMcpServers}
+            className="border-none text-theme-text-secondary hover:text-cta-button flex items-center gap-x-1"
+          >
+            <ArrowClockwise
+              size={16}
+              className={loadingMcpServers ? "animate-spin" : ""}
+            />
+            <p className="text-sm">
+              {loadingMcpServers ? "Loading..." : "Refresh"}
+            </p>
+          </button>
+        </div>
+      </div>
+      {children({ loadingMcpServers })}
+    </>
+  );
+}
+
+export function MCPServersList({
+  isLoading = false,
   servers = [],
   selectedServer,
   handleClick,
 }) {
+  if (isLoading) {
+    return (
+      <div className="text-theme-text-secondary text-center text-xs flex flex-col gap-y-2">
+        <p>Loading MCP Servers from configuration file...</p>
+        <a
+          href="https://docs.anythingllm.com/mcp-servers/getting-started"
+          target="_blank"
+          className="text-theme-text-secondary underline hover:text-cta-button"
+        >
+          Learn more about MCP Servers.
+        </a>
+      </div>
+    );
+  }
+
   if (servers.length === 0) {
     return (
       <div className="text-theme-text-secondary text-center text-xs flex flex-col gap-y-2">
@@ -22,7 +113,7 @@ export default function MCPServersList({
   }
 
   return (
-    <div className="bg-theme-bg-secondary text-white rounded-xl min-w-[360px] w-fit">
+    <div className="bg-theme-bg-secondary text-white rounded-xl w-full md:min-w-[360px]">
       {servers.map((server, index) => (
         <div
           key={server.name}
